+++ src/ast/passes/field_analyser.h
@@ -34,7 +34,7 @@ public:
   void visit(ArrayAccess &arr) override;
   void visit(Cast &cast) override;
   void visit(Sizeof &szof) override;
-  void visit(Offsetof &ofof) override;
+  void visit(Offsetof &offof) override;
   void visit(AssignMapStatement &assignment) override;
   void visit(AssignVarStatement &assignment) override;
   void visit(Unop &unop) override;