+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- `offsetof()` now supports sub fields e.g. `offsetof(struct Foo, bar.a.b);`
+  - [#3761](https://github.com/bpftrace/bpftrace/pull/3761)
 - Pointers may now be used in if conditions, tenary conditions and as operands in logical AND and OR expressions
   - [#3656](https://github.com/bpftrace/bpftrace/pull/3656)
 #### Changed